sample-python-django-app/requirements.txt
Django==1.9
django-dashing
whitenoise